chore: all types and props are inside a respective file now
