--- core/types/@types/ThemeContextValue-type.ts
@@ -1,6 +1,0 @@
-export type ThemeContextValue = {
-  appTheme: "light" | "dark";
-  toggleTheme: () => void;
-  setAppThemeToLight: () => void;
-  setAppThemeToDark: () => void;
-};
