--- core/types/@types/index.ts
@@ -1,3 +1,0 @@
-export * from "./selectOptions-type";
-export * from "./ThemeContextValue-type";
-export * from "./user-type";
