--- core/types/@types/selectOptions-type.ts
@@ -1,4 +1,0 @@
-export type SelectOptions = {
-  value: string;
-  text: string;
-};
