--- core/types/@types/user-type.ts
@@ -1,5 +1,0 @@
-export type UserType = {
-  id: string;
-  name: string | null;
-  avatar: string | null;
-};
