+++ core/types/index.ts
@@ -1,2 +1,2 @@
-export * from "./@types";
 export * from "./props";
+export * from "./types";
