--- core/types/props/AuthContext-props.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from "react";
-
-export interface AuthProviderProps {
-  children: ReactNode | ReactNode[];
-}
