--- core/types/props/AvatarPopover-props.ts
@@ -1,6 +1,0 @@
-import { UserType } from "../@types";
-
-export interface AvatarPopoverProps {
-  user: UserType;
-  className?: string;
-}
