--- core/types/props/Button-props.ts
@@ -1,6 +1,0 @@
-import { ButtonHTMLAttributes, ReactNode } from "react";
-
-export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  ring?: boolean;
-  children: ReactNode | ReactNode[];
-}
