--- core/types/props/Illustration-props.ts
@@ -1,3 +1,0 @@
-import { HTMLAttributes } from "react";
-
-export interface IllustrationProps extends HTMLAttributes<HTMLDivElement> {}
