--- core/types/props/PageContainer-props.ts
@@ -1,8 +1,0 @@
-import { ReactNode } from "react";
-
-export interface PageContainerProps {
-  headTitle?: string;
-  description?: string;
-  center?: boolean;
-  children: ReactNode | ReactNode[];
-}
