--- core/types/props/PageLayout-props.ts
@@ -1,7 +1,0 @@
-import { ReactNode } from "react";
-
-export interface PageLayoutProps {
-  navbar?: boolean;
-  footer?: boolean;
-  children: ReactNode | ReactNode[];
-}
