--- core/types/props/Select-props.ts
@@ -1,7 +1,0 @@
-import { SelectOptions } from "../@types";
-
-export interface SelectProps {
-  selectedValue: string;
-  changeValue: (value: any) => void;
-  options: SelectOptions[];
-}
