--- core/types/props/SignInList-props.ts
@@ -1,5 +1,0 @@
-import { AuthContext } from "@core/contexts";
-
-export interface SignInListProps {
-  providers: { signInWithGoogle: AuthContext["signInWithGoogle"] };
-}
