--- core/types/props/ThemeContext-props.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from "react";
-
-export interface ThemeProviderProps {
-  children: ReactNode | ReactNode[];
-}
