--- core/types/props/index.ts
@@ -1,9 +1,0 @@
-export * from "./AuthContext-props";
-export * from "./AvatarPopover-props";
-export * from "./Button-props";
-export * from "./Illustration-props";
-export * from "./PageContainer-props";
-export * from "./PageLayout-props";
-export * from "./Select-props";
-export * from "./SignInList-props";
-export * from "./ThemeContext-props";
